chore: resolve fetch models api limit rate
