+++ plugins/model-management-plugin/@types/global.d.ts
@@ -1,2 +1,3 @@
 declare const PLUGIN_NAME: string;
 declare const MODULE_PATH: string;
+declare const MODEL_CATALOG_URL: string;
