+++ plugins/model-management-plugin/tsconfig.json
@@ -1,7 +1,7 @@
 {
   "compilerOptions": {
     "target": "es2016",
-    "module": "ES6",
+    "module": "esnext",
     "moduleResolution": "node",
     "outDir": "./dist",
     "esModuleInterop": true,
