+++ plugins/model-management-plugin/webpack.config.js
@@ -19,6 +19,9 @@ module.exports = {
     new webpack.DefinePlugin({
       PLUGIN_NAME: JSON.stringify(packageJson.name),
       MODULE_PATH: JSON.stringify(`${packageJson.name}/${packageJson.module}`),
+      MODEL_CATALOG_URL: JSON.stringify(
+        "https://cdn.jsdelivr.net/gh/janhq/models@0.1.1/src/dist/index.js"
+      ),
     }),
   ],
   output: {
@@ -30,7 +33,7 @@ module.exports = {
     extensions: [".ts", ".js"],
   },
   optimization: {
-    minimize: false
+    minimize: false,
   },
   // Add loaders and other configuration as needed for your project
 };
